refactor: code review #53 - fix correctness, security, and code quality issues
Code Review Findings & Fixes:

Critical:
- Remove meaningless std::isfinite() checks on int64_t amounts in
  protocol.cpp (int64_t is always finite; the checks were dead code)
- Fix constant_time_compare() timing leak: eliminated early return on
  size mismatch that allowed length oracle attacks via timing side-channel
- Add missing #include <nlohmann/json.hpp> in metrics_collector.hpp
  (previously relied on transitive includes)
- Fix check_suspicious_activity() never working: last_activity_ was never
  initialized or updated; message_count_ was incremented but never checked
  properly

Security:
- Use static_cast<unsigned char> for isalnum/isxdigit/iscntrl/isspace
  calls in protocol_validation.cpp to avoid undefined behavior with
  negative char values
- Replace is_valid_session_id substr() allocation with compare() to avoid
  unnecessary heap allocation in validation hot path

Code Quality:
- Remove duplicate thread safety comment block in websocket_session.hpp
- Move error_code_to_string into cppsim::server::detail namespace for
  proper encapsulation while maintaining testability
- Fix session_metrics.hpp doc comment mismatch (increment methods
  incorrectly labeled as "Get total bytes")
- Fix validate_config() redundant checks that re-validated constraints
  already enforced by load_from_json() per-field clamping
- Make load_from_env() env_mappings map static const to avoid redundant
  construction on each call
- Fix validate_message_envelope() catch clause to use std::exception&
  instead of nlohmann::json::exception& for broader exception coverage
- Fix error_templates.hpp format_error<> which was fundamentally broken
  (variadic template expansion cannot iterate packs sequentially as coded)
- Initialize last_activity_ in websocket_session constructor

All 77 tests pass.

+++ src/server/session_metrics.hpp
@@ -58,12 +58,12 @@ public:
      * @param time Processing time to record
      */
     void update_processing_time(std::chrono::microseconds time) noexcept {
-        total_processing_time_.fetch_add(time.count(), std::memory_order_relaxed);
+        total_processing_time_.fetch_add(static_cast<uint64_t>(time.count()), std::memory_order_relaxed);
         processing_time_count_.fetch_add(1, std::memory_order_relaxed);
         
         // Update max processing time
         uint64_t current_max = max_processing_time_.load(std::memory_order_relaxed);
-        while (time.count() > current_max) {
+        while (static_cast<uint64_t>(time.count()) > current_max) {
             if (max_processing_time_.compare_exchange_weak(
                     current_max, 
                     static_cast<uint64_t>(time.count()),
@@ -79,12 +79,12 @@ public:
      * @param duration Connection duration
      */
     void update_connection_duration(std::chrono::seconds duration) noexcept {
-        total_connection_time_.fetch_add(duration.count(), std::memory_order_relaxed);
+        total_connection_time_.fetch_add(static_cast<uint64_t>(duration.count()), std::memory_order_relaxed);
         connection_count_.fetch_add(1, std::memory_order_relaxed);
         
         // Update max connection time
         uint64_t current_max = max_connection_time_.load(std::memory_order_relaxed);
-        while (duration.count() > current_max) {
+        while (static_cast<uint64_t>(duration.count()) > current_max) {
             if (max_connection_time_.compare_exchange_weak(
                     current_max, 
                     static_cast<uint64_t>(duration.count()),
@@ -135,7 +135,7 @@ public:
         if (count == 0) return 0.0;
         
         uint64_t total = total_processing_time_.load(std::memory_order_relaxed);
-        return static_cast<double>(total) / count / 1000.0; // Convert microseconds to milliseconds
+        return static_cast<double>(total) / static_cast<double>(count) / 1000.0; // Convert microseconds to milliseconds
     }
     
     /**
@@ -155,7 +155,7 @@ public:
         if (count == 0) return 0.0;
         
         uint64_t total = total_connection_time_.load(std::memory_order_relaxed);
-        return static_cast<double>(total) / count;
+        return static_cast<double>(total) / static_cast<double>(count);
     }
     
     /**
@@ -175,16 +175,16 @@ public:
     }
     
     /**
-     * @brief Get total bytes sent
-     * @return Total number of bytes sent
+     * @brief Increment bytes sent counter
+     * @param bytes Number of bytes to increment by
      */
     void increment_bytes_sent(size_t bytes) noexcept {
         bytes_sent_.fetch_add(bytes, std::memory_order_relaxed);
     }
     
     /**
-     * @brief Get total bytes received
-     * @return Total number of bytes received
+     * @brief Increment bytes received counter
+     * @param bytes Number of bytes to increment by
      */
     void increment_bytes_received(size_t bytes) noexcept {
         bytes_received_.fetch_add(bytes, std::memory_order_relaxed);
@@ -336,7 +336,7 @@ public:
                << ", received=" << total_bytes_received_
                << " | Rate limits: " << total_rate_limit_exceeded_
                << " | Avg per session: " << (session_count_ > 0 ? 
-                   static_cast<double>(total_messages_sent_) / session_count_ : 0.0) << " msgs sent";
+                   static_cast<double>(total_messages_sent_) / static_cast<double>(session_count_) : 0.0) << " msgs sent";
             return ss.str();
         } catch (...) {
             return "Server metrics unavailable";
@@ -352,7 +352,7 @@ public:
         if (session_count_ == 0) return true;  // No sessions is fine
         
         double error_rate = static_cast<double>(total_errors_) / 
-                           (total_messages_sent_ + total_messages_received_);
+                           (static_cast<double>(total_messages_sent_) + static_cast<double>(total_messages_received_));
         
         // Consider unhealthy if error rate is above 5% or too many errors
         return error_rate < 0.05 && total_errors_ < 1000;
@@ -393,10 +393,10 @@ private:
     void update_averages() noexcept {
         if (session_count_ > 0) {
             uint64_t total_messages = total_messages_sent_ + total_messages_received_;
-            avg_messages_per_session_ = static_cast<double>(total_messages) / session_count_;
+            avg_messages_per_session_ = static_cast<double>(total_messages) / static_cast<double>(session_count_);
             
             if (total_messages > 0) {
-                avg_error_rate_ = static_cast<double>(total_errors_) / total_messages;
+                avg_error_rate_ = static_cast<double>(total_errors_) / static_cast<double>(total_messages);
             } else {
                 avg_error_rate_ = 0.0;
             }
